requirements/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.1
MarkupSafe==1.0
marshmallow==3.0.0b8
psycopg2==2.7.4
requests==2.18.4
SQLAlchemy==1.2.4
urllib3==1.22
Werkzeug==0.14.1
redis==2.10.6
pycrypto==2.6.1